% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypso_bautzen.R
\docType{data}
\name{hypso_bautzen}
\alias{hypso_bautzen}
\title{Hypsographic Table of Bautzen Reservoir}
\format{
A data frame or tibble with at least two columns:
\describe{
\item{level}{Numeric, the water level in meters.}
\item{area}{Numeric, the surface area in square kilometers or hectares.}
\item{volume}{Numeric, the water volume in cubic meters or million cubic meters.}
}
}
\source{
The Limnology Department at TU Dresden.
}
\description{
A dataset providing the hypsographic table for the Bautzen Reservoir.
This table relates the water level of the reservoir to its surface area and volume,
which is crucial for limnological and hydrological studies.
}
\references{
\url{http://www.tu-dresden.de/hydrobiologie}
}
\keyword{data}
