% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klimphyt-package.R
\docType{package}
\name{klimphyt-package}
\alias{klimphyt}
\alias{klimphyt-package}
\title{klimphyt: Helper Functions for the KlimPhyt Project}
\description{
Support data management and visualisation of phytoplankton time series.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tpetzoldt/klimphyt/}
}

}
\author{
\strong{Maintainer}: Thomas Petzoldt \email{thomas.petzoldt@tu-dresden.de} (\href{https://orcid.org/0000-0002-4951-6468}{ORCID}) (042aqky30)

}
\keyword{internal}
