% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyto_aggregate_groups.R
\name{phyto_aggregate_groups}
\alias{phyto_aggregate_groups}
\title{Aggregate Phytoplankton Data by Group}
\usage{
phyto_aggregate_groups(phyto, fill_missing = TRUE)
}
\arguments{
\item{phyto}{A data frame containing phytoplankton data. The function requires
the following columns:
\itemize{
\item \strong{date}: A column representing the date of observation.
\item \strong{depth}: A column representing the sampling depth.
\item \strong{group}: A column for the phytoplankton group/species name.
\item \strong{bv}: A column with the biovolume or abundance value to be aggregated.
}}

\item{fill_missing}{Logical value, if \code{TRUE} (default) missing values
in the vertical profile are set to zero to avoid weighting bias.
This is especially important for rarer species.}
}
\value{
A data frame with the aggregated biovolume (\code{bv}) for each
\code{date}, \code{depth}, and \code{group}.
}
\description{
This function aggregates phytoplankton data by \code{group}, ensuring that all
depths present for a given date are accounted for, even if a specific group
has a missing value at that depth. It fills these missing values with zero
before aggregation to avoid introducing bias in the depth-weighted mean.
}
